统计详情页实现、优化存储结构

+++ app/src/main/kotlin/indi/dmzz_yyhyy/lightnovelreader/ui/book/content/ContentScreen.kt
@@ -257,7 +257,7 @@ fun Content(
         while (isRunning) {
             val now = LocalTime.now()
             val elapsed = 1 /*Duration.between(lastUpdate, now).seconds.coerceAtLeast(1)*/
-            accumulateReadingTime(bookId, elapsed.toInt())
+            accumulateReadingTime(bookId, elapsed)
             lastUpdate = now
             delay(1000)
         }
